+++ src/components/panels/edit/modals/SubjectEditor.vue
@@ -154,6 +154,11 @@
                       :style="`${this.preferenceStore.styleModalTextColor()}`">
                       {{ type.label }}</li>
                   </ol>
+
+                  <template v-if="showTypes">
+                    <input id="provisionalCheck" type="checkbox" v-model="provisionalHeading">
+                    <label for="provisionalCheck"  style="word-wrap:break-word">Provisional</label>
+                  </template>
                 </div>
                 <div style="flex:1">
 
@@ -838,6 +843,7 @@ export default {
         // 'madsrdf:GenreForm': { label: 'Genre ($v)', value: 'madsrdf:GenreForm', selected: false },
         'madsrdf:Geographic': { label: 'Geographic ($z)', value: 'madsrdf:Geographic', selected: false },
         'madsrdf:Temporal': { label: 'Chronological ($y)', value: 'madsrdf:Temporal', selected: false },
+        'bf:Family': { label: 'Family ($a)', value: 'bf:Family', selected: false },                      // madsrdf:family name doesn't work with the conversion
       },
 
       labelMap: {
@@ -876,7 +882,9 @@ export default {
         "sources", "sees", "lcclasses", "lcclasss", "identifiers", "broaders",
         "collections", "subjects", "marcKeys"
       ],
-      selectedSortOrder: ""
+      selectedSortOrder: "",
+
+      provisionalHeading: false,
 
     }
   },
@@ -2574,10 +2582,9 @@ export default {
       if (allHaveURI && allHaveType) {
         this.okayToAdd = true
       }
-      if (allHaveURI && !allHaveType && this.components.length == 1) {
-        this.okayToAdd = true
-      }
-
+      // if (allHaveURI && !allHaveType && this.components.length > 1) { // why was this here?
+      //   this.okayToAdd = true
+      // }
 
 
     },
@@ -2979,7 +2986,13 @@ export default {
         this.components = newComponents
       }
 
-      console.info("components: ", this.components)
+      if (this.provisionalHeading){
+        for (let comp of this.components){
+          if (!comp.uri){
+            comp.provisional = true
+          }
+        }
+      }
 
       this.$emit('subjectAdded', this.components)
     },
@@ -3024,7 +3037,8 @@ export default {
 
       this.contextData = { nodeMap: {} }
       this.authorityLookupLocal = null,
-        this.subjectString = ''
+      this.subjectString = ''
+      this.provisionalHeading = false
 
     },