+++ tests-playwright/components/complex_lookup/subject_builder.spec.js
@@ -585,6 +585,29 @@ test('Add a literal and set type Temporal, it has the correct XML', async ({ pag
     await expect(page.locator('#app')).toContainText('literal');
 });
 
+test('Add a literal and set type Family, it has the correct XML', async ({ page }) => {
+    await page.goto('http://localhost:5555/marva/');
+
+    // Update the preferences for this test
+    let prefs = JSON.stringify(preferences)
+    await page.evaluate(prefs => localStorage.setItem("marva-preferences", prefs), prefs)
+    await page.reload();
+
+    await page.getByText('Click Here').click();
+    await page.getByRole('button', { name: 'Monograph', exact: true }).nth(1).click();
+
+    await page.locator('form').filter({ hasText: 'Search LCSH/LCNAF' }).getByRole('textbox').click();
+    await page.locator('form').filter({ hasText: 'Search LCSH/LCNAFbolt' }).getByRole('textbox').fill('l');
+    await page.getByRole('textbox', { name: 'Enter Subject Headings Here' }).fill('literal');
+    await page.getByText('literal [Literal]').click();
+    await page.getByText('Family ($a)').click();
+    await page.getByRole('checkbox', { name: 'Provisional' }).check();
+    await page.getByRole('button', { name: 'Add [SHIFT+Enter]' }).click();
+    await page.getByText('bf:Work').click();
+    await expect(page.locator('#app')).toContainText('<bf:Family>');
+    await expect(page.locator('#app')).toContainText('literal');
+});
+
 test('Add a Dogs with literal topic sub it has the correct XML', async ({ page }) => {
     await page.goto('http://localhost:5555/marva/');