+++ CHANGELOG.md
@@ -1,18 +1,33 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
-## [1.6.3] - 2025-06-26
+## [1.6.4] - 2026-07-24
+### Update
+- Add "bf:Family" to subject literal types
+- `$7` appears at the end of a NAR
+- Prefer "Hang" for "Hani" when no match for "kore"
+- Support work type picklist for profiles
+- Remove scriptshifter from case changes
+
+## [1.6.4] - 2026-07-17
+### Update
+- Adjust NAR creation
+  - Don't create empty fields
+  - stricter indicator check
+  - Don't allow post if field data or subfield are empty
+
+## [1.6.3] - 2026-06-26
 ### Added
 - Buttons for `Create NAR/Hub` when looking for a hub
 
 ### Fixed
 - Sometimes after posting a record and then opening a unposted record, the `Post` button would be "green," this shouldn't happen anymore
 
-## [1.6.2] - 2025-04-28
+## [1.6.2] - 2026-04-28
 ### Update
 - Copycat has a MARC comparison view. The comparison is shown after selecting a profile to load with and provides a chance for the cataloger to backout. Comparison can also be envoked from a "Compare" button if an existing record is found.
 
-## [1.6.1] - 2025-04-27
+## [1.6.1] - 2026-04-27
 ### Update
 - 046 generation in NAR creation should be more smoothly updating on change
 - When multiple LCCs are set for a work and use the Tool nav bar to open shelflisting it will only use the first LCC component
@@ -22,7 +37,7 @@ All notable changes to this project will be documented in this file.
 
 
 
-## [1.6.x] - 2025-03-20
+## [1.6.x] - 2026-03-20
 ### Added
 - Introduced MLC Number Generator (https://bibframe.org/docs/view/documentation-marva-manual/Marva%20tools/mlc-numbers.md)